import directly from patched file
